fstab: only for reference: add more mount options
Add more mount options for the data partition:
[1] eMMC    - the default
[2] SD-card
[3] mSATA
Options 2, 3 are commented out. They are present for reference,
and for simplification of a switch if required so.

Notice,
while 'system' and 'cache' partitions can also be moved, this might
be of little profit, as ...
'system' partition is read only, and thus its size is no limitation
'cache' partition is used primarily for OTA updates, which are not
	supported on CM-QS600.